+++ Battle_interface/sea/battle_shipcommand.h
@@ -30,7 +30,7 @@ protected:
 	long m_nChargeTextureNum;
 	long m_nCommandTextureNum;
 
-	std::vector<long> m_aChargeQuantity;
+	std::vector<int32_t> m_aChargeQuantity;
 };
 
 #endif
